Fixes the context menu code
The context menu is created new based on the item in the card history
that is selected. This code setups of events to check which context menu
button is pressed to know which card was pressed. There were issues
with the code to determine that card name. Those are fixed in this
commit.

Removes commented out code

Instead of cloning the divider and menu item element they are instead
created with copy + pasted styles from SpellTable running at this
date/time. This ensures that the context menu adding code works when
tokens and other cards with a limited context menu do not appear.

The context menu code also checks to make sure that only one context
menu item is added for "Cast to Spellslinger"